Upgrade OpenTelemetry components from v0.122.0 to v0.125.0

+++ internal/component/otelcol/config_queue.go
@@ -12,10 +12,10 @@ import (
 // QueueArguments holds shared settings for components which can queue
 // requests.
 type QueueArguments struct {
-	Enabled      bool `alloy:"enabled,attr,optional"`
-	NumConsumers int  `alloy:"num_consumers,attr,optional"`
-	QueueSize    int  `alloy:"queue_size,attr,optional"`
-	Blocking     bool `alloy:"blocking,attr,optional"`
+	Enabled      bool  `alloy:"enabled,attr,optional"`
+	NumConsumers int   `alloy:"num_consumers,attr,optional"`
+	QueueSize    int64 `alloy:"queue_size,attr,optional"`
+	Blocking     bool  `alloy:"blocking,attr,optional"`
 
 	// Storage is a binding to an otelcol.storage.* component extension which handles
 	// reading and writing to disk
